pyproject.toml
[project]
name = "python-analytic"
version = "0.1.0"
description = ""
authors = [
    {name = "Andrey Novichenko",email = "60508206+anovicenko74@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib (>=3.10.3,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "aiogram (>=3.20.0.post0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
